+++ src/people/widgetViews/__tests__/OrganizationDetails.spec.tsx
@@ -8,7 +8,6 @@ import nock from 'nock';
 import React from 'react';
 import { Organization, mainStore } from 'store/main';
 import OrganizationDetails from '../OrganizationDetails';
-import AddUserModal from '../organization/AddUserModal';
 
 jest.mock('react-router-dom', () => ({
   ...jest.requireActual('react-router-dom'),
@@ -37,21 +36,21 @@ const organization: Organization = {
  */
 describe('OrganizationDetails', () => {
   jest.setTimeout(20000);
-  nock.disableNetConnect();
+  //nock.disableNetConnect();
   nock(user.url).get('/person/id/1').reply(200, { user });
   nock(user.url).get('/ask').reply(200, {});
 
   it('Add User button shows Add User modal', async () => {
     const props: {
       close: jest.Mock;
-      org: Organization | undefined; 
+      org: Organization | undefined;
       resetOrg: jest.Mock;
       getOrganizations: jest.Mock;
       openAddUserModal: jest.Mock;
     } = {
       close: jest.fn(),
       org: {
-        id: "51",
+        id: '51',
         name: 'TEST_NEW',
         owner_pubkey: '03cbb9c01cdcf91a3ac3b543a556fbec9c4c3c2a6ed753e19f2706012a26367ae3',
         uuid: 'cmas9gatu2rvqiev4ur0',
@@ -62,29 +61,29 @@ describe('OrganizationDetails', () => {
         github: '',
         img: '',
         website: '',
-        show: false, 
+        show: false
       },
       resetOrg: jest.fn(),
       getOrganizations: jest.fn(),
-      openAddUserModal: jest.fn(),
+      openAddUserModal: jest.fn()
     };
-  
+
     const spyOpenAddUserModal = jest.spyOn(props, 'openAddUserModal');
-  
+
     const { getByText } = render(
       <Router>
         <OrganizationDetails {...props} />
-      </Router>   
-      );
-  
+      </Router>
+    );
+
     (async () => {
       await waitFor(() => {
         const addUserButton = getByText('Add User');
-      expect(addUserButton).toBeInTheDocument();
-      fireEvent.click(addUserButton);
-      expect(spyOpenAddUserModal).toHaveBeenCalledTimes(1);
-      })    
-    })(); 
+        expect(addUserButton).toBeInTheDocument();
+        fireEvent.click(addUserButton);
+        expect(spyOpenAddUserModal).toHaveBeenCalledTimes(1);
+      });
+    })();
   });
   it('render organization name', async () => {
     jest.spyOn(mainStore, 'getPersonAssignedBounties').mockReturnValue(Promise.resolve([]));
@@ -261,43 +260,5 @@ describe('OrganizationDetails', () => {
       expect(withdrawBtn).toBeDisabled();
       expect(historyBtn).toBeDisabled();
     });
-  });
-});
-
-describe('AddUserModal Component', () => {
-  it('filters people according to the search term, enables "Add User" Button and adds user when "Add User" button is clicked', async () => {
-    const mockProps = {
-      isOpen: true,
-      close: jest.fn(),
-      loading: false,
-      onSubmit: jest.fn(),
-      disableFormButtons: false,
-      setDisableFormButtons: jest.fn(),
-    };
-
-    const mockUser = {
-      owner_pubkey: '03cbb9c01cdcf91a3ac3b543a556fbec9c4c3c2a6ed753e19f2706012a26367ae3',
-      owner_alias: 'Ednum',
-      img: '',
-    };
-
-    const { getByPlaceholderText, getByText } = render(
-      <AddUserModal {...mockProps} />
-    );
-
-    (async () => {
-      const searchInput = getByPlaceholderText('Type to search ...');
-      fireEvent.change(searchInput, { target: { value: 'Anish' } });
-      await waitFor(() => expect(getByText('Anish Yadav')).toBeInTheDocument());
-      await waitFor(() => {
-        const addUserButton = getByText('Add User');
-        expect(addUserButton).toBeEnabled();
-        fireEvent.click(addUserButton);
-        expect(mockProps.onSubmit).toHaveBeenCalledTimes(1);
-        expect(mockProps.onSubmit).toHaveBeenCalledWith({
-          owner_pubkey: mockUser.owner_pubkey,
-        });
-      });
-    })();   
   });
 });
